ISSUE #5 変更: GithubRepository データクラスを model ディレクトリに移動しました
