+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/CustomAdapter.kt
@@ -7,6 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.model.GithubRepository
 
 class CustomAdapter(
     private val itemClickListener: OnItemClickListener,
