+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/model/GithubRepository.kt
@@ -1,0 +1,15 @@
+package jp.co.yumemi.android.code_check.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class GithubRepository(
+    val name: String,
+    val ownerIconUrl: String,
+    val language: String,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val forksCount: Long,
+    val openIssuesCount: Long,
+) : Parcelable
