+++ app/src/main/res/navigation/nav_graph.xml
@@ -22,7 +22,7 @@
         tools:layout="@layout/fragment_search_results_detail_screen">
         <argument
             android:name="githubRepository"
-            app:argType="jp.co.yumemi.android.code_check.GithubRepository" />
+            app:argType="jp.co.yumemi.android.code_check.model.GithubRepository" />
     </fragment>
 
 </navigation>
